OC_flask/requirements.txt
Flask==2.2.2
Flask_WTF==1.0.1
matplotlib==3.5.2
numpy==1.23.0
opencv_python==4.6.0.66
Pillow==9.2.0
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
Werkzeug==2.2.2
WTForms==3.0.1
